Refactor on_update to redude nesting to a max of 1
As the nesting in the on_update methode was way to deep to be able to
properly read the methode a refactor was necessary. To achive the
nesting reduction, the fact that the match construct is also an
expression with a value was used. The successful path just returns the
valid object while the error path immedialty return the on update
function passing on the error to the caller.